// packages/snap/src/features/homepage/types.tsx
export enum HomePageNames {
  RemoveCustodianToken = 'homepage-remove-custodian-token',
  CancelRemoveToken = 'homepage-cancel-remove-token',
  RemoveToken = 'homepage-remove-token',
  CancelToken = 'homepage-cancel-token',
}

export enum HomePagePrefixes {
  SelectCustodian = 'homepage-select-',
  ConnectToken = 'homepage-connect-token-',
}